export refactor; submit styling
