+++ src/pages/index/index.styles.ts
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+import { Button } from "components/common";
+
+export const Submit = styled(Button)`
+  margin: 10px;
+`;
