+++ src/store/index.ts
@@ -20,3 +20,4 @@ export function setupStore(preloadedState?: PreloadedState<RootState>) {
 export type RootState = ReturnType<typeof rootReducer>;
 export type AppStore = ReturnType<typeof setupStore>;
 export type AppDispatch = AppStore["dispatch"];
+export * from "./types";
