+++ src/store/sagas/personSaga.ts
@@ -1,7 +1,7 @@
 import { PayloadAction } from "@reduxjs/toolkit";
 import { put, takeLatest, all } from "redux-saga/effects";
 import { setPersonAction } from "store/features/personSlice";
-import { SET_PERSON } from "store/types";
+import { SET_PERSON } from "store";
 import { Person } from "types";
 
 export function* setPerson(action: PayloadAction<Person>) {
